継承を辞める
